Using Gradle 3.0 for building means that we had to use a newer version of the TestKit API. standardOutput nwo has to be forwardOutput() to a local StringWriter.
