+++ jruby-gradle-jar-plugin/src/test/groovy/com/github/jrubygradle/jar/JRubyJarSpec.groovy
@@ -18,7 +18,6 @@ class JRubyJarSpec extends Specification {
         project = ProjectBuilder.builder().build()
         project.gradle.startParameter.offline = true
         project.with {
-            logging.level = LIFECYCLE
             apply plugin: 'com.github.jruby-gradle.base'
             jruby.defaultRepositories = false
         }
